Migrate permissions to new sourcing module (#3353)
* Migrate permissions to new sourcing module

+++ delta/app/src/test/scala/ch/epfl/bluebrain/nexus/delta/utils/RouteFixtures.scala
@@ -159,7 +159,7 @@ trait RouteFixtures extends TestHelpers with IOValues {
     )
 
   def permissionsMetadata(
-      rev: Long = 1L,
+      rev: Int = 1,
       deprecated: Boolean = false,
       createdBy: Subject = Anonymous,
       updatedBy: Subject = Anonymous
